+++ src/main/java/uk/ac/soton/comp2300/group42/energyclient/ui/model/ApplianceModel.java
@@ -18,7 +18,7 @@ public class ApplianceModel {
         return dto;
     }
 
-    public int getId() { return dto.getId(); }
+    public Long getId() { return dto.getId(); }
 
     public String getName() { return name.get(); }
     public void setName(String value) { name.set(value); }
